* New win32 buildsystem. * VVector alternative is not included in this change -- waiting review.
git-svn-id: https://ecell.svn.sourceforge.net/svnroot/ecell/ecell3/trunk@2729 f1531174-cb10-0410-9fe6-89aa7ac3eedb

+++ ecell/dm/ODEStepper.cpp
@@ -493,7 +493,7 @@ bool ODEStepper::calculateRadauIIA()
 	  if ( theta < 0.99 )
 	    {
 	      eta = theta / ( 1.0 - theta );
-	      const Real anIterationError( eta * aNorm * pow( theta, getMaxIterationNumber() - 2 - anIterator ) / theStoppingCriterion );
+	      const Real anIterationError( eta * aNorm * pow( theta, static_cast<int>(getMaxIterationNumber() - 2 - anIterator) ) / theStoppingCriterion );
 	      
 	      if ( anIterationError >= 1.0 )
 		{
